Refactor non & ndn to support new source and rmeta acl, etc

+++ src/component/cyfs-stack/src/ndn_api/acl/ndn_inner.rs
@@ -1,4 +1,3 @@
-use crate::acl::*;
 use crate::ndn::*;
 use cyfs_base::*;
 use cyfs_lib::*;
@@ -6,31 +5,30 @@ use cyfs_lib::*;
 use std::sync::Arc;
 
 pub(crate) struct NDNAclInnerInputProcessor {
-    acl: AclManagerRef,
     next: NDNInputProcessorRef,
 }
 
 impl NDNAclInnerInputProcessor {
-    pub fn new(acl: AclManagerRef, next: NDNInputProcessorRef) -> NDNInputProcessorRef {
-        let ret = Self { acl, next };
+    pub fn new(next: NDNInputProcessorRef) -> NDNInputProcessorRef {
+        let ret = Self { next };
         Arc::new(Box::new(ret))
+    }
+
+    fn check_access(&self, service: &str, common: &NDNInputRequestCommon) -> BuckyResult<()> {
+        common.source.check_current_zone(service)
     }
 }
 
 #[async_trait::async_trait]
 impl NDNInputProcessor for NDNAclInnerInputProcessor {
     async fn put_data(&self, req: NDNPutDataInputRequest) -> BuckyResult<NDNPutDataInputResponse> {
-        self.acl
-            .check_local_zone_permit("ndn in put_data", &req.common.source)
-            .await?;
+        self.check_access("ndn.put_data", &req.common)?;
 
         self.next.put_data(req).await
     }
 
     async fn get_data(&self, req: NDNGetDataInputRequest) -> BuckyResult<NDNGetDataInputResponse> {
-        self.acl
-        .check_local_zone_permit("ndn in get_data", &req.common.source)
-        .await?;
+        self.check_access("ndn.get_data", &req.common)?;
 
         self.next.get_data(req).await
     }
@@ -39,9 +37,7 @@ impl NDNInputProcessor for NDNAclInnerInputProcessor {
         &self,
         req: NDNDeleteDataInputRequest,
     ) -> BuckyResult<NDNDeleteDataInputResponse> {
-        self.acl
-            .check_local_zone_permit("ndn in delete_data", &req.common.source)
-            .await?;
+        self.check_access("ndn.delete_data", &req.common)?;
 
         self.next.delete_data(req).await
     }
@@ -50,64 +46,7 @@ impl NDNInputProcessor for NDNAclInnerInputProcessor {
         &self,
         req: NDNQueryFileInputRequest,
     ) -> BuckyResult<NDNQueryFileInputResponse> {
-        self.acl
-            .check_local_zone_permit("ndn in query_file", &req.common.source)
-            .await?;
-
-        self.next.query_file(req).await
-    }
-}
-
-// ndn的内部使用input processor作为output
-pub(crate) struct NDNAclInnerOutputProcessor {
-    acl: AclManagerRef,
-    target: DeviceId,
-    next: NDNInputProcessorRef,
-}
-
-impl NDNAclInnerOutputProcessor {
-    pub fn new(acl: AclManagerRef, target: DeviceId, next: NDNInputProcessorRef) -> NDNInputProcessorRef {
-        let ret = Self { acl, target, next };
-        Arc::new(Box::new(ret))
-    }
-}
-
-#[async_trait::async_trait]
-impl NDNInputProcessor for NDNAclInnerOutputProcessor {
-    async fn put_data(&self, req: NDNPutDataInputRequest) -> BuckyResult<NDNPutDataInputResponse> {
-        self.acl
-            .check_local_zone_permit("ndn out put_data", &self.target)
-            .await?;
-
-        self.next.put_data(req).await
-    }
-
-    async fn get_data(&self, req: NDNGetDataInputRequest) -> BuckyResult<NDNGetDataInputResponse> {
-        self.acl
-        .check_local_zone_permit("ndn out get_data", &self.target)
-        .await?;
-
-        self.next.get_data(req).await
-    }
-
-    async fn delete_data(
-        &self,
-        req: NDNDeleteDataInputRequest,
-    ) -> BuckyResult<NDNDeleteDataInputResponse> {
-        self.acl
-            .check_local_zone_permit("ndn out delete_data", &self.target)
-            .await?;
-
-        self.next.delete_data(req).await
-    }
-
-    async fn query_file(
-        &self,
-        req: NDNQueryFileInputRequest,
-    ) -> BuckyResult<NDNQueryFileInputResponse> {
-        self.acl
-            .check_local_zone_permit("ndn out query_file", &self.target)
-            .await?;
+        self.check_access("ndn.query_file", &req.common)?;
 
         self.next.query_file(req).await
     }
